이홍준_Update_Damage
몬스터 접촉공격, 몬스터 원거리공격시 체력 잃음 구현, 체력바 연동

+++ Assets/Scripts/Entities/CharacterStats.cs
@@ -14,8 +14,8 @@ public enum StatsChangeType
 public class CharacterStats : MonoBehaviour
 {
     public StatsChangeType statsChangeType;
-    [Range(1, 100)] public int maxHealth;
-    [Range(1, 100)] public int currentHealth;
+    [Range(1, 100)] public float maxHealth;
+    [Range(1, 100)] public float currentHealth;
     [Range(1f, 20f)] public float speed;
     [Range(1f, 100f)] public float exp;
     public int level = 1;
